pyproject.toml
[project]
name = "ask-your-documents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "chromadb>=1.5.8",
    "openai>=2.33.0",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "streamlit>=1.57.0",
    "voyageai>=0.3.7",
]
